+++ library/src/main/java/cloud/pace/sdk/appkit/app/deeplink/RedirectUriReceiverActivity.kt
@@ -1,4 +1,4 @@
-package cloud.pace.sdk.appkit.app.customtab
+package cloud.pace.sdk.appkit.app.deeplink
 
 import android.content.Intent
 import android.os.Bundle
@@ -9,8 +9,8 @@ class RedirectUriReceiverActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        // Handling the redirect in this way ensures that we can remove the browser custom tab from the back stack
-        val newIntent = Intent(this, CustomTabManagementActivity::class.java)
+        // Handling the redirect in this way ensures that we can remove the WebViewActivity or custom tab from the back stack
+        val newIntent = Intent(this, DeepLinkManagementActivity::class.java)
         newIntent.data = intent.data
         newIntent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_SINGLE_TOP)
         startActivity(newIntent)